requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
langchain-huggingface==0.0.3
langchain-community==0.3.7
transformers==4.44.2
faiss-cpu==1.8.0
numpy==2.1.2
gunicorn==23.0.0
waitress==3.0.0
redis==5.0.1